TM-436 cannot save deleted shared resource

+++ common/common-kotlin/src/main/kotlin/com/testerum/common_kotlin/path_extensions.kt
@@ -99,6 +99,10 @@ fun JavaPath.isNotSameFileAs(other: JavaPath): Boolean = !this.isSameFileAs(othe
  */
 fun JavaPath.smartMoveTo(other: JavaPath,
                          createDestinationExistsException: () -> Exception) {
+    if (this.doesNotExist) {
+        return
+    }
+
     if (other.doesNotExist) {
         other.parent?.createDirectories()
         Files.move(this, other)
